# backend/docs/swagger.yaml
basePath: /api
definitions:
  main.Booking:
    properties:
      dateTime:
        type: string
      employeeName:
        type: string
      facilityName:
        type: string
      id:
        type: integer
      status:
        description: Confirmed, Pending, Completed, Cancelled
        type: string
    type: object
  main.createBookingInput:
    properties:
      dateTime:
        type: string
      employeeName:
        type: string
      facilityName:
        type: string
      status:
        type: string
    required:
    - dateTime
    - employeeName
    - facilityName
    type: object
  main.statsResponse:
    properties:
      activeFacilities:
        type: integer
      pendingRequests:
        type: integer
      registeredUsers:
        type: integer
      totalBookings:
        type: integer
    type: object
  main.updateBookingInput:
    properties:
      dateTime:
        type: string
      employeeName:
        type: string
      facilityName:
        type: string
      status:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: API for managing forest campus facility bookings
  title: Forest Facility Management API
  version: "1.0"
paths:
  /bookings:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Booking'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all bookings
      tags:
      - bookings
    post:
      consumes:
      - application/json
      parameters:
      - description: Booking data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.createBookingInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Booking'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new booking
      tags:
      - bookings
  /bookings/{id}:
    delete:
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a booking
      tags:
      - bookings
    get:
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Booking'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get booking by ID
      tags:
      - bookings
    put:
      consumes:
      - application/json
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.updateBookingInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Booking'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a booking
      tags:
      - bookings
  /health:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - health
  /stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.statsResponse'
      summary: Dashboard statistics
      tags:
      - stats
swagger: "2.0"